# docs/swagger.yaml
basePath: /
definitions:
...
  types.ErrorResponse:
    properties:
      details:
        description: Additional error details
      error:
        description: Error code/type
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  types.JobStatusResponse:
    properties:
      episode_id:
        description: Episode ID
        type: integer
      error:
        description: Error message (only for failed status)
        type: string
      error_code:
        description: Specific error code like "403", "timeout", "corrupt_file" (only
          for failed jobs)
        type: string
      error_details:
        description: Technical error details for debugging (only for failed jobs)
        type: string
      error_type:
        description: 'Error type: "download", "processing", "system" (only for failed
          jobs)'
        type: string
      hint:
        description: Helpful hint for the client (e.g., "Use retry=true parameter")
        type: string
      job_id:
        description: Job ID (optional)
        type: integer
      max_retries:
        description: Maximum retry attempts (only for failed jobs)
        type: integer
      message:
        description: Human-readable message
        type: string
      progress:
        description: Progress 0-100
        type: integer
      retried:
        description: True if this was a manual retry (only when applicable)
        type: boolean
      retry_after:
        description: Seconds until retry (only for failed jobs)
        type: number
      retry_count:
        description: Number of retries attempted (only for failed jobs)
        type: integer
      status:
        description: 'Status: pending, processing, completed, failed, permanently_failed,
          not_found'
        type: string
    type: object
...
  types.TranscriptionData:
    properties:
      cached:
        description: Whether data is cached - optional for some responses
        type: boolean
      duration:
        description: Duration in seconds
        example: 300.5
        type: number
      episode_id:
        description: Podcast Index Episode ID (optional for some responses)
        type: integer
      language:
        description: Detected or specified language
        example: en
        type: string
      model:
        description: Model used for transcription
        example: ggml-base.en.bin
        type: string
      source:
        description: '"fetched" or "generated" - optional for some responses'
        type: string
      text:
        description: Full transcription text
        example: This is the transcription...
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: Podcast Player API Support
    url: https://github.com/killallgit/killallplayer-api
  description: |-
    A REST API for podcast discovery and episode management.
    This API provides endpoints for searching podcasts, managing episodes,
    and generating waveforms for audio visualization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Podcast Player API
  version: "1.0"
paths:
...
  /api/v1/episodes/{id}/transcribe:
    get:
      consumes:
      - application/json
      description: |-
        Retrieve the full transcription text for a podcast episode if available. Transcriptions may come
        from two sources: 'fetched' (downloaded from podcast RSS feed transcriptURL) or 'generated' (created
        using Whisper speech-to-text). The response includes the full text, source type, language, and timestamps.
        Use POST /episodes/{id}/transcribe first to trigger generation if transcription doesn't exist.
      parameters:
      - description: Episode's Podcast Index ID
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Full transcription text with metadata
          schema:
            $ref: '#/definitions/types.TranscriptionData'
        "400":
          description: Invalid episode ID format
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: No transcription available for this episode
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Database or service error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get episode transcription text
      tags:
      - transcription
    post:
      consumes:
      - application/json
      description: |-
        Trigger transcription for a podcast episode. The system first checks if a transcript is available
        at the episode's transcriptURL (from RSS feed). If found, it fetches and stores it. Otherwise, if
        Whisper is configured, it generates a transcription using speech-to-text. Transcription is an async
        process that may take several minutes depending on episode duration. Poll the status endpoint with job_id to track progress.
      parameters:
      - description: Episode's Podcast Index ID
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transcription already exists and is ready
          schema:
            $ref: '#/definitions/types.JobStatusResponse'
        "202":
          description: Transcription job queued successfully (use job_id to track)
          schema:
            $ref: '#/definitions/types.JobStatusResponse'
        "400":
          description: Invalid episode ID format
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Service unavailable or configuration error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Generate or fetch episode transcription
      tags:
      - transcription
  /api/v1/episodes/{id}/transcribe/status:
    get:
      consumes:
      - application/json
      description: Check the status of transcription generation for an episode
      parameters:
      - description: Episode ID (Podcast Index ID)
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transcription status
          schema:
            $ref: '#/definitions/types.JobStatusResponse'
        "400":
          description: Invalid Podcast Index Episode ID
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Transcription not available
          schema:
            $ref: '#/definitions/types.JobStatusResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get transcription generation status
      tags:
      - transcription
...
schemes:
- http
- https
swagger: "2.0"
...